feat(home): add usage limit and promotional banners, refactor Home component for improved layout
